refactor: split into request::score module
