+++ src/request/mod.rs
@@ -28,6 +28,7 @@ mod matches;
 mod news;
 mod ranking;
 mod replay;
+mod score;
 mod seasonal_backgrounds;
 mod serialize;
 mod user;
@@ -42,6 +43,7 @@ pub use matches::*;
 pub use news::*;
 pub use ranking::*;
 pub use replay::*;
+pub use score::*;
 pub use seasonal_backgrounds::*;
 pub use user::*;
 pub use wiki::*;
